removing old test infrastructure incl testing module + updated toml and readme
