--- openmsitoolbox/testing/__init__.py
@@ -1,9 +1,0 @@
-" Imports available from openmsitoolbox.testing "
-
-from .test_with_logger import TestWithLogger
-from .test_with_output_location import TestWithOutputLocation
-
-__all__ = [
-    "TestWithLogger",
-    "TestWithOutputLocation",
-]
