+++ test/README.md
@@ -1,0 +1,16 @@
+## Development & Testing
+
+This project uses **Poetry** for dependency management.
+
+### Setup
+Install dependencies (including test dependencies):
+
+```bash
+poetry install --with test
+```
+
+then
+
+```bash
+poetry run pytest
+```
