+++ test/test_scripts/test_controlled_process.py
@@ -1,21 +1,13 @@
-" Tests for the 'ControlledProcess' classes "
-
-# imports
-import unittest
 import time
+import pytest
 from openmsitoolbox.utilities.exception_tracking_thread import ExceptionTrackingThread
 from openmsitoolbox import ControlledProcessSingleThread, ControlledProcessMultiThreaded
 
-# some constants
 TIMEOUT_SECS = 10
 N_THREADS = 3
 
 
 class ControlledProcessSingleThreadForTesting(ControlledProcessSingleThread):
-    """
-    Class to use in testing ControlledProcessSingleThread
-    """
-
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.counter = 0
@@ -34,10 +26,6 @@ class ControlledProcessSingleThreadForTesting(ControlledProcessSingleThread):
 
 
 class ControlledProcessMultiThreadedForTesting(ControlledProcessMultiThreaded):
-    """
-    Class to use in test ControlledProcessMultiThreaded
-    """
-
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.counter = 0
@@ -58,94 +46,50 @@ class ControlledProcessMultiThreadedForTesting(ControlledProcessMultiThreaded):
                     self.counter += 1
 
 
-class TestControlledProcess(unittest.TestCase):
-    """
-    Class for testing ControlledProcess utility classes
-    """
+@pytest.mark.parametrize(
+    "ProcessClass,kwargs",
+    [
+        (ControlledProcessSingleThreadForTesting, {"update_secs": 5}),
+        (ControlledProcessMultiThreadedForTesting, {"n_threads": N_THREADS, "update_secs": 5}),
+    ],
+)
+def test_controlled_process(ProcessClass, kwargs):
+    """Test both single- and multi-threaded controlled process variants."""
+    cp = ProcessClass(**kwargs)
+    assert cp.counter == 0
 
-    def test_controlled_process_single_thread(self):
-        """
-        Test the single-thread controlled process
-        """
-        cpst = ControlledProcessSingleThreadForTesting(update_secs=5)
-        self.assertEqual(cpst.counter, 0)
-        run_thread = ExceptionTrackingThread(target=cpst.run)
-        run_thread.start()
-        try:
-            self.assertFalse(cpst.checked)
-            time.sleep(1.0)
-            cpst.control_command_queue.put("c")
-            cpst.control_command_queue.put("check")
-            time.sleep(1.0)
-            self.assertTrue(cpst.checked)
-            self.assertFalse(cpst.on_shutdown_called)
-            cpst.control_command_queue.put("q")
-            time.sleep(2.0)
-            self.assertTrue(cpst.on_shutdown_called)
-            run_thread.join(timeout=TIMEOUT_SECS)
-            time.sleep(2.0)
-            if run_thread.is_alive():
-                errmsg = (
-                    "ERROR: running thread in test_controlled_process_single_thread "
-                    f"timed out after {TIMEOUT_SECS} seconds!"
-                )
-                raise TimeoutError(errmsg)
-            self.assertEqual(cpst.counter, 5)
-        except Exception as exc:
-            raise exc
-        finally:
-            if run_thread.is_alive():
-                try:
-                    cpst.shutdown()
-                    run_thread.join(timeout=5)
-                    if run_thread.is_alive():
-                        errmsg = (
-                            "ERROR: running thread in test_controlled_process_single_thread "
-                            "timed out after 5 seconds!"
-                        )
-                        raise TimeoutError(errmsg)
-                except Exception as exc:
-                    raise exc
+    run_thread = ExceptionTrackingThread(target=cp.run)
+    run_thread.start()
 
-    def test_controlled_process_multi_threaded(self):
-        """
-        Test the multi-threaded controlled process
-        """
-        cpmt = ControlledProcessMultiThreadedForTesting(
-            n_threads=N_THREADS, update_secs=5
-        )
-        self.assertEqual(cpmt.counter, 0)
-        run_thread = ExceptionTrackingThread(target=cpmt.run)
-        run_thread.start()
-        try:
-            self.assertFalse(cpmt.checked)
-            time.sleep(0.5)
-            cpmt.control_command_queue.put("c")
-            cpmt.control_command_queue.put("check")
-            time.sleep(0.5)
-            self.assertTrue(cpmt.checked)
-            self.assertFalse(cpmt.on_shutdown_called)
-            cpmt.control_command_queue.put("q")
-            time.sleep(1.0)
-            self.assertTrue(cpmt.on_shutdown_called)
-            run_thread.join(timeout=TIMEOUT_SECS)
+    try:
+        # --- pre-check ---
+        assert not cp.checked
+
+        # --- trigger commands ---
+        time.sleep(1.0)
+        cp.control_command_queue.put("c")
+        cp.control_command_queue.put("check")
+        time.sleep(1.0)
+
+        # --- assertions ---
+        assert cp.checked
+        assert not cp.on_shutdown_called
+
+        # --- shutdown ---
+        cp.control_command_queue.put("q")
+        time.sleep(2.0)
+
+        assert cp.on_shutdown_called
+        run_thread.join(timeout=TIMEOUT_SECS)
+
+        if run_thread.is_alive():
+            pytest.fail(f"Thread timed out after {TIMEOUT_SECS}s")
+
+        assert cp.counter == 5
+
+    finally:
+        if run_thread.is_alive():
+            cp.shutdown()
+            run_thread.join(timeout=5)
             if run_thread.is_alive():
-                errmsg = (
-                    "ERROR: running thread in test_controlled_process_multi_threaded "
-                    f"timed out after {TIMEOUT_SECS} seconds!"
-                )
-                raise TimeoutError(errmsg)
-            self.assertEqual(cpmt.counter, 5)
-        except Exception as exc:
-            raise exc
-        finally:
-            if run_thread.is_alive():
-                try:
-                    cpmt.shutdown()
-                    run_thread.join(timeout=5)
-                    if run_thread.is_alive():
-                        raise TimeoutError(
-                            "ERROR: running thread timed out after 5 seconds!"
-                        )
-                except Exception as exc:
-                    raise exc
+                pytest.fail("Thread did not terminate after forced shutdown")
